requirements.txt
# 本地文档解析（文上传 PDF/Word/TXT）依赖
PyPDF2>=3.0.0
python-docx>=1.0.0
